requirements.txt
numpy
pyarrow
torch
pyyaml
scipy
tensorboard
pandas
scikit-learn
matplotlib